+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
-## 1.27.1-2 (In development)
+## 1.27.1-3 (In development)
+
+Based on [apache/commons-compress@d0aec98](https://github.com/apache/commons-compress/commit/d0aec98f62d1870d22f869aeec5ac1ed9ea3126a).
+
+## 1.27.1-2 (2026-03-25)
 
 Based on [apache/commons-compress@d0aec98](https://github.com/apache/commons-compress/commit/d0aec98f62d1870d22f869aeec5ac1ed9ea3126a).